pyproject.toml
[tool.poetry]
name = "ams"
version = "0.1.0"
description = ""
authors = ["PvtJoker91 <sanyakoenig@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
django-filter = "^23.5"
djoser = "^2.2.2"
django-silk = "^5.1.0"
django-simple-history = "^3.4.0"
drf-spectacular = "^0.27.1"
django-cors-headers = "^4.3.1"
celery = {extras = ["redis"], version = "^5.3.6"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
